+++ packages/crdt_socket_sync/test/src/utils/mock_handler.dart
@@ -1,3 +1,6 @@
+import 'dart:convert';
+import 'dart:typed_data';
+
 import 'package:crdt_lf/crdt_lf.dart';
 
 import 'mock_operation.dart';
@@ -10,7 +13,7 @@ class MockHandler extends Handler<String> {
   String get id => 'test-handler';
 
   @override
-  String getSnapshotState() => 'test_state';
+  Uint8List getSnapshotState() => Uint8List.fromList(utf8.encode('test_state'));
 
   @override
   OperationFactory get operationFactory => (payload) => MockOperation(this);